# source code/practice/api-doc.json
{"openapi":"3.1.0","info":{"title":"OpenAPI definition","version":"v0"},"servers":[{"url":"http://localhost:8080","description":"Generated server url"}],"paths":{"/direct-chats/{partnerId}":{"put":{"tags":["chat-controller"],"operationId":"getChat","parameters":[{"name":"partnerId","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"202":{"description":"Accepted","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ChatResponse"}}}},"409":{"description":"Conflict","content":{"application/problem+json":{"schema":{"$ref":"#/components/schemas/ProblemDetail"}}}},"404":{"description":"Not Found","content":{"application/problem+json":{"schema":{"$ref":"#/components/schemas/ProblemDetail"}}}},"400":{"description":"Bad Request","content":{"application/problem+json":{"schema":{"$ref":"#/components/schemas/ProblemDetail"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/DirectChatResponse"}}}}}}},"/chats/{chatId}/texts/{postingId}":{"put":{"tags":["posting-controller"],"operationId":"createText","parameters":[{"name":"postingId","in":"path","required":true,"schema":{"type":"string","format":"uuid"}},{"name":"chatId","in":"path","required":true,"schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/TextRequest"}}},"required":true},"responses":{"409":{"description":"Conflict","content":{"application/problem+json":{"schema":{"$ref":"#/components/schemas/ProblemDetail"}}}},"404":{"description":"Not Found","content":{"application/problem+json":{"schema":{"$ref":"#/components/schemas/ProblemDetail"}}}},"400":{"description":"Bad Request","content":{"application/problem+json":{"schema":{"$ref":"#/components/schemas/ProblemDetail"}}}},"202":{"description":"Accepted","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PostingResponse"}}}}}}},"/chats/{chatId}/seens/{postingId}":{"put":{"tags":["posting-controller"],"operationId":"createSeen","parameters":[{"name":"chatId","in":"path","required":true,"schema":{"type":"string"}},{"name":"postingId","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SeenRequest"}}},"required":true},"responses":{"409":{"description":"Conflict","content":{"application/problem+json":{"schema":{"$ref":"#/components/schemas/ProblemDetail"}}}},"404":{"description":"Not Found","content":{"application/problem+json":{"schema":{"$ref":"#/components/schemas/ProblemDetail"}}}},"400":{"description":"Bad Request","content":{"application/problem+json":{"schema":{"$ref":"#/components/schemas/ProblemDetail"}}}},"202":{"description":"Accepted","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PostingResponse"}}}}}}},"/chats/{chatId}/images/{postingId}":{"put":{"tags":["posting-controller"],"operationId":"createImage","parameters":[{"name":"chatId","in":"path","required":true,"schema":{"type":"string"}},{"name":"postingId","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ImageRequest"}}},"required":true},"responses":{"409":{"description":"Conflict","content":{"application/problem+json":{"schema":{"$ref":"#/components/schemas/ProblemDetail"}}}},"404":{"description":"Not Found","content":{"application/problem+json":{"schema":{"$ref":"#/components/schemas/ProblemDetail"}}}},"400":{"description":"Bad Request","content":{"application/problem+json":{"schema":{"$ref":"#/components/schemas/ProblemDetail"}}}},"202":{"description":"Accepted","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PostingResponse"}}}}}}},"/chats/{chatId}/icons/{postingId}":{"put":{"tags":["posting-controller"],"operationId":"createIcon","parameters":[{"name":"chatId","in":"path","required":true,"schema":{"type":"string"}},{"name":"postingId","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/IconRequest"}}},"required":true},"responses":{"409":{"description":"Conflict","content":{"application/problem+json":{"schema":{"$ref":"#/components/schemas/ProblemDetail"}}}},"404":{"description":"Not Found","content":{"application/problem+json":{"schema":{"$ref":"#/components/schemas/ProblemDetail"}}}},"400":{"description":"Bad Request","content":{"application/problem+json":{"schema":{"$ref":"#/components/schemas/ProblemDetail"}}}},"202":{"description":"Accepted","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PostingResponse"}}}}}}},"/chats/{chatId}/files/{postingId}":{"put":{"tags":["posting-controller"],"operationId":"createFile","parameters":[{"name":"chatId","in":"path","required":true,"schema":{"type":"string"}},{"name":"postingId","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/FileRequest"}}},"required":true},"responses":{"409":{"description":"Conflict","content":{"application/problem+json":{"schema":{"$ref":"#/components/schemas/ProblemDetail"}}}},"404":{"description":"Not Found","content":{"application/problem+json":{"schema":{"$ref":"#/components/schemas/ProblemDetail"}}}},"400":{"description":"Bad Request","content":{"application/problem+json":{"schema":{"$ref":"#/components/schemas/ProblemDetail"}}}},"202":{"description":"Accepted","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PostingResponse"}}}}}}},"/users":{"post":{"tags":["user-controller"],"operationId":"registerUser","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SignUpRequest"}}},"required":true},"responses":{"400":{"description":"Bad Request","content":{"application/problem+json":{"schema":{"$ref":"#/components/schemas/ProblemDetail"}}}},"409":{"description":"Conflict","content":{"application/problem+json":{"schema":{"$ref":"#/components/schemas/ProblemDetail"}}}},"404":{"description":"Not Found","content":{"application/problem+json":{"schema":{"$ref":"#/components/schemas/ProblemDetail"}}}},"201":{"description":"Created","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ProfileResponse"}}}}}}},"/tokens/refresh":{"post":{"tags":["token-controller"],"operationId":"refresh","responses":{"409":{"description":"Conflict","content":{"application/problem+json":{"schema":{"$ref":"#/components/schemas/ProblemDetail"}}}},"404":{"description":"Not Found","content":{"application/problem+json":{"schema":{"$ref":"#/components/schemas/ProblemDetail"}}}},"400":{"description":"Bad Request","content":{"application/problem+json":{"schema":{"$ref":"#/components/schemas/ProblemDetail"}}}},"200":{"description":"OK"}}}},"/tokens/oauth2":{"post":{"tags":["token-controller"],"operationId":"exchange","responses":{"409":{"description":"Conflict","content":{"application/problem+json":{"schema":{"$ref":"#/components/schemas/ProblemDetail"}}}},"404":{"description":"Not Found","content":{"application/problem+json":{"schema":{"$ref":"#/components/schemas/ProblemDetail"}}}},"400":{"description":"Bad Request","content":{"application/problem+json":{"schema":{"$ref":"#/components/schemas/ProblemDetail"}}}},"200":{"description":"OK"}}}},"/profiles/me/password":{"post":{"tags":["profile-controller"],"operationId":"changePassword","parameters":[{"name":"password","in":"query","required":false,"schema":{"type":"string"}},{"name":"new_password","in":"query","required":true,"schema":{"type":"string"}}],"responses":{"409":{"description":"Conflict","content":{"application/problem+json":{"schema":{"$ref":"#/components/schemas/ProblemDetail"}}}},"404":{"description":"Not Found","content":{"application/problem+json":{"schema":{"$ref":"#/components/schemas/ProblemDetail"}}}},"400":{"description":"Bad Request","content":{"application/problem+json":{"schema":{"$ref":"#/components/schemas/ProblemDetail"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/AccountResponse"}}}}}}},"/groups":{"post":{"tags":["chat-controller"],"operationId":"createGroupChat","parameters":[{"name":"partnerId","in":"query","required":true,"schema":{"type":"array","items":{"type":"string","format":"uuid"},"maxItems":50,"minItems":2,"uniqueItems":true}},{"name":"roomName","in":"query","required":true,"schema":{"type":"string"}}],"responses":{"202":{"description":"Accepted","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ChatResponse"}}}},"409":{"description":"Conflict","content":{"application/problem+json":{"schema":{"$ref":"#/components/schemas/ProblemDetail"}}}},"404":{"description":"Not Found","content":{"application/problem+json":{"schema":{"$ref":"#/components/schemas/ProblemDetail"}}}},"400":{"description":"Bad Request","content":{"application/problem+json":{"schema":{"$ref":"#/components/schemas/ProblemDetail"}}}},"201":{"description":"Created","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ChatResponse"}}}}}}},"/files/upload-urls":{"post":{"tags":["presign-controller"],"operationId":"uploadUrl","parameters":[{"name":"filename","in":"query","required":true,"schema":{"type":"string"}}],"responses":{"409":{"description":"Conflict","content":{"application/problem+json":{"schema":{"$ref":"#/components/schemas/ProblemDetail"}}}},"404":{"description":"Not Found","content":{"application/problem+json":{"schema":{"$ref":"#/components/schemas/ProblemDetail"}}}},"400":{"description":"Bad Request","content":{"application/problem+json":{"schema":{"$ref":"#/components/schemas/ProblemDetail"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/S3PresignedResponse"}}}}}}},"/profiles/me":{"get":{"tags":["profile-controller"],"operationId":"get","responses":{"409":{"description":"Conflict","content":{"application/problem+json":{"schema":{"$ref":"#/components/schemas/ProblemDetail"}}}},"404":{"description":"Not Found","content":{"application/problem+json":{"schema":{"$ref":"#/components/schemas/ProblemDetail"}}}},"400":{"description":"Bad Request","content":{"application/problem+json":{"schema":{"$ref":"#/components/schemas/ProblemDetail"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ProfileResponse"}}}}}},"patch":{"tags":["profile-controller"],"operationId":"changeProfile","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProfileRequest"}}},"required":true},"responses":{"409":{"description":"Conflict","content":{"application/problem+json":{"schema":{"$ref":"#/components/schemas/ProblemDetail"}}}},"404":{"description":"Not Found","content":{"application/problem+json":{"schema":{"$ref":"#/components/schemas/ProblemDetail"}}}},"400":{"description":"Bad Request","content":{"application/problem+json":{"schema":{"$ref":"#/components/schemas/ProblemDetail"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ProfileResponse"}}}}}}},"/chats/{identifier}/preference":{"patch":{"tags":["settings-controller"],"operationId":"setPreference","parameters":[{"name":"identifier","in":"path","required":true,"schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PreferenceRequest"}}},"required":true},"responses":{"409":{"description":"Conflict","content":{"application/problem+json":{"schema":{"$ref":"#/components/schemas/ProblemDetail"}}}},"404":{"description":"Not Found","content":{"application/problem+json":{"schema":{"$ref":"#/components/schemas/ProblemDetail"}}}},"400":{"description":"Bad Request","content":{"application/problem+json":{"schema":{"$ref":"#/components/schemas/ProblemDetail"}}}},"202":{"description":"Accepted"}}}},"/users/me/logs":{"get":{"tags":["log-controller"],"operationId":"listLog","parameters":[{"name":"anchorSequenceNumber","in":"query","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"409":{"description":"Conflict","content":{"application/problem+json":{"schema":{"$ref":"#/components/schemas/ProblemDetail"}}}},"404":{"description":"Not Found","content":{"application/problem+json":{"schema":{"$ref":"#/components/schemas/ProblemDetail"}}}},"400":{"description":"Bad Request","content":{"application/problem+json":{"schema":{"$ref":"#/components/schemas/ProblemDetail"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/InboxLogResponse"}}}}}}}},"/user-infos":{"get":{"tags":["use-info-controller"],"operationId":"getUsersByIds","parameters":[{"name":"userId","in":"query","required":true,"schema":{"type":"array","items":{"type":"string","format":"uuid"},"maxItems":50,"minItems":0,"uniqueItems":true}}],"responses":{"409":{"description":"Conflict","content":{"application/problem+json":{"schema":{"$ref":"#/components/schemas/ProblemDetail"}}}},"404":{"description":"Not Found","content":{"application/problem+json":{"schema":{"$ref":"#/components/schemas/ProblemDetail"}}}},"400":{"description":"Bad Request","content":{"application/problem+json":{"schema":{"$ref":"#/components/schemas/ProblemDetail"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object","additionalProperties":{"$ref":"#/components/schemas/UserInfo"}}}}}}}},"/themes":{"get":{"tags":["theme-controller"],"operationId":"getThemes","responses":{"409":{"description":"Conflict","content":{"application/problem+json":{"schema":{"$ref":"#/components/schemas/ProblemDetail"}}}},"404":{"description":"Not Found","content":{"application/problem+json":{"schema":{"$ref":"#/components/schemas/ProblemDetail"}}}},"400":{"description":"Bad Request","content":{"application/problem+json":{"schema":{"$ref":"#/components/schemas/ProblemDetail"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/ThemeResponse"}}}}}}}},"/presences":{"get":{"tags":["presence-controller"],"operationId":"find","parameters":[{"name":"chatId","in":"query","required":true,"schema":{"type":"array","items":{"type":"string"},"maxItems":50,"minItems":1,"uniqueItems":true}}],"responses":{"409":{"description":"Conflict","content":{"application/problem+json":{"schema":{"$ref":"#/components/schemas/ProblemDetail"}}}},"404":{"description":"Not Found","content":{"application/problem+json":{"schema":{"$ref":"#/components/schemas/ProblemDetail"}}}},"400":{"description":"Bad Request","content":{"application/problem+json":{"schema":{"$ref":"#/components/schemas/ProblemDetail"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object","additionalProperties":{"$ref":"#/components/schemas/RoomPresenceResponse"}}}}}}}},"/people":{"get":{"tags":["search-controller"],"summary":"Search people by name, keywords","operationId":"findUsers","parameters":[{"name":"query","in":"query","required":true,"schema":{"type":"string"}}],"responses":{"409":{"description":"Conflict","content":{"application/problem+json":{"schema":{"$ref":"#/components/schemas/ProblemDetail"}}}},"404":{"description":"Not Found","content":{"application/problem+json":{"schema":{"$ref":"#/components/schemas/ProblemDetail"}}}},"400":{"description":"Bad Request","content":{"application/problem+json":{"schema":{"$ref":"#/components/schemas/ProblemDetail"}}}},"200":{"description":"List of matched users","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/PeopleResponse"}}}}}},"security":[{"bearerAuth":[]}]}},"/conversations":{"get":{"tags":["conversation-controller"],"operationId":"list","parameters":[{"name":"anchorRevisionNumber","in":"query","required":false,"schema":{"type":"integer","format":"int64"}}],"responses":{"409":{"description":"Conflict","content":{"application/problem+json":{"schema":{"$ref":"#/components/schemas/ProblemDetail"}}}},"404":{"description":"Not Found","content":{"application/problem+json":{"schema":{"$ref":"#/components/schemas/ProblemDetail"}}}},"400":{"description":"Bad Request","content":{"application/problem+json":{"schema":{"$ref":"#/components/schemas/ProblemDetail"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/ConversationResponse"}}}}}}}},"/chats/{chatId}":{"get":{"tags":["chat-controller"],"operationId":"getChat_1","parameters":[{"name":"chatId","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"202":{"description":"Accepted","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ChatResponse"}}}},"409":{"description":"Conflict","content":{"application/problem+json":{"schema":{"$ref":"#/components/schemas/ProblemDetail"}}}},"404":{"description":"Not Found","content":{"application/problem+json":{"schema":{"$ref":"#/components/schemas/ProblemDetail"}}}},"400":{"description":"Bad Request","content":{"application/problem+json":{"schema":{"$ref":"#/components/schemas/ProblemDetail"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ChatResponse"}}}}}}},"/chats/{chatId}/messages":{"get":{"tags":["message-controller"],"operationId":"listMessages","parameters":[{"name":"chatId","in":"path","required":true,"schema":{"type":"string"}},{"name":"anchorSequenceNumber","in":"query","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"409":{"description":"Conflict","content":{"application/problem+json":{"schema":{"$ref":"#/components/schemas/ProblemDetail"}}}},"404":{"description":"Not Found","content":{"application/problem+json":{"schema":{"$ref":"#/components/schemas/ProblemDetail"}}}},"400":{"description":"Bad Request","content":{"application/problem+json":{"schema":{"$ref":"#/components/schemas/ProblemDetail"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"oneOf":[{"$ref":"#/components/schemas/MessageStateResponse"},{"$ref":"#/components/schemas/FileStateResponse"},{"$ref":"#/components/schemas/IconStateResponse"},{"$ref":"#/components/schemas/ImageStateResponse"},{"$ref":"#/components/schemas/PreferenceStateResponse"},{"$ref":"#/components/schemas/TextStateResponse"}]}}}}}}}},"/chats/{chatId}/logs":{"get":{"tags":["log-controller"],"operationId":"listLog_1","parameters":[{"name":"chatId","in":"path","required":true,"schema":{"type":"string"}},{"name":"anchorSequenceNumber","in":"query","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"409":{"description":"Conflict","content":{"application/problem+json":{"schema":{"$ref":"#/components/schemas/ProblemDetail"}}}},"404":{"description":"Not Found","content":{"application/problem+json":{"schema":{"$ref":"#/components/schemas/ProblemDetail"}}}},"400":{"description":"Bad Request","content":{"application/problem+json":{"schema":{"$ref":"#/components/schemas/ProblemDetail"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/InboxLogResponse"}}}}}}}},"/chat-histories/{chatId}":{"get":{"tags":["search-controller"],"summary":"Find messages in a chat","operationId":"findMessages","parameters":[{"name":"chatId","in":"path","required":true,"schema":{"type":"string"}},{"name":"query","in":"query","required":true,"schema":{"type":"string"}}],"responses":{"409":{"description":"Conflict","content":{"application/problem+json":{"schema":{"$ref":"#/components/schemas/ProblemDetail"}}}},"404":{"description":"Chat not found","content":{"application/problem+json":{"schema":{"$ref":"#/components/schemas/ProblemDetail"},"example":null}}},"400":{"description":"Bad Request","content":{"application/problem+json":{"schema":{"$ref":"#/components/schemas/ProblemDetail"}}}},"200":{"description":"List of matched messages","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/MessageResponse"}}}}}},"security":[{"bearerAuth":[]}]}},"/buddy-presences":{"get":{"tags":["presence-controller"],"operationId":"listBuddies","responses":{"409":{"description":"Conflict","content":{"application/problem+json":{"schema":{"$ref":"#/components/schemas/ProblemDetail"}}}},"404":{"description":"Not Found","content":{"application/problem+json":{"schema":{"$ref":"#/components/schemas/ProblemDetail"}}}},"400":{"description":"Bad Request","content":{"application/problem+json":{"schema":{"$ref":"#/components/schemas/ProblemDetail"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/BuddyResponse"}}}}}}}}},"components":{"schemas":{"ProblemDetail":{"type":"object","properties":{"type":{"type":"string","format":"uri"},"title":{"type":"string"},"status":{"type":"integer","format":"int32"},"detail":{"type":"string"},"instance":{"type":"string","format":"uri"},"properties":{"type":"object","additionalProperties":{}}}},"ChatResponse":{"type":"object","properties":{"identifier":{"type":"string"},"maxParticipants":{"type":"integer","format":"int32"},"creators":{"type":"array","items":{"type":"string","format":"uuid"},"uniqueItems":true},"preference":{"$ref":"#/components/schemas/PreferenceResponse"}}},"PreferenceResponse":{"type":"object","properties":{"iconId":{"type":"integer","format":"int32"},"customName":{"type":"string"},"customAvatar":{"type":"string"},"themeBackground":{"type":"string"},"themeName":{"type":"string"}}},"DirectChatResponse":{"type":"object","properties":{"mapping":{"$ref":"#/components/schemas/DirectMapping"},"newly":{"type":"boolean"}}},"DirectMapping":{"type":"object","properties":{"caller":{"type":"string","format":"uuid"},"partner":{"type":"string","format":"uuid"},"chatId":{"type":"string"}}},"TextRequest":{"type":"object","properties":{"content":{"type":"string","minLength":1}},"required":["content"]},"PostingResponse":{"type":"object","properties":{"postingId":{"type":"string","format":"uuid"},"senderId":{"type":"string","format":"uuid"},"chatId":{"type":"string"},"createdAt":{"type":"string","format":"date-time"}}},"SeenRequest":{"type":"object","properties":{"at":{"type":"string","format":"date-time"}},"required":["at"]},"ImageRequest":{"type":"object","properties":{"uri":{"type":"string","minLength":1},"filename":{"type":"string","minLength":1},"width":{"type":"integer","format":"int32","maximum":2000,"minimum":200},"height":{"type":"integer","format":"int32","maximum":2000,"minimum":200},"format":{"type":"string"}},"required":["filename","uri"]},"IconRequest":{"type":"object","properties":{"iconId":{"type":"integer","format":"int32"}},"required":["iconId"]},"FileRequest":{"type":"object","properties":{"filename":{"type":"string"},"size":{"type":"integer","format":"int32"},"uri":{"type":"string"}},"required":["filename","size","uri"]},"SignUpRequest":{"type":"object","properties":{"username":{"type":"string","maxLength":20,"minLength":5,"pattern":"\\S+"},"password":{"type":"string"},"name":{"type":"string","minLength":1},"gender":{"type":"number","format":"float"},"dob":{"type":"string","format":"date-time"},"avatar":{"type":"string"}},"required":["gender","name","username"]},"ProfileResponse":{"type":"object","properties":{"id":{"type":"string","format":"uuid"},"username":{"type":"string"},"name":{"type":"string"},"dob":{"type":"string","format":"date-time"},"role":{"type":"string"},"avatar":{"type":"string"},"gender":{"type":"string"}}},"AccessToken":{"type":"object","properties":{"accessToken":{"type":"string"},"refreshToken":{"type":"string"}}},"AccountResponse":{"type":"object","properties":{"profile":{"$ref":"#/components/schemas/ProfileResponse"},"accessToken":{"$ref":"#/components/schemas/AccessToken"}}},"S3PresignedResponse":{"type":"object","properties":{"key":{"type":"string"},"bucket":{"type":"string"},"filename":{"type":"string"},"presignedUploadUrl":{"type":"string"},"downloadUrl":{"type":"string"}}},"ProfileRequest":{"type":"object","properties":{"name":{"type":"string","pattern":"\\S.*"},"gender":{"type":"number","format":"float"},"dob":{"type":"string","format":"date-time"},"avatar":{"type":"string"}}},"PreferenceRequest":{"type":"object","properties":{"iconId":{"type":"integer","format":"int32"},"customName":{"type":"string"},"themeId":{"type":"integer","format":"int64"},"customAvatar":{"type":"string"}}},"FileStateResponse":{"allOf":[{"$ref":"#/components/schemas/MessageStateResponse"},{"type":"object","properties":{"filename":{"type":"string"},"size":{"type":"integer","format":"int32"},"uri":{"type":"string"}}}]},"IconStateResponse":{"allOf":[{"$ref":"#/components/schemas/MessageStateResponse"},{"type":"object","properties":{"iconId":{"type":"integer","format":"int32"}}}]},"ImageSpecResponse":{"type":"object","properties":{"uri":{"type":"string"},"filename":{"type":"string"},"width":{"type":"integer","format":"int32"},"height":{"type":"integer","format":"int32"},"format":{"type":"string"}}},"ImageStateResponse":{"allOf":[{"$ref":"#/components/schemas/MessageStateResponse"},{"type":"object","properties":{"image":{"$ref":"#/components/schemas/ImageSpecResponse"}}}]},"InboxLogResponse":{"type":"object","properties":{"sequenceNumber":{"type":"integer","format":"int64"},"chatId":{"type":"string"},"roomName":{"type":"string"},"roomAvatar":{"type":"string"},"revisionNumber":{"type":"integer","format":"int64"},"sender":{"$ref":"#/components/schemas/PartnerResponse"},"ownerId":{"type":"string","format":"uuid"},"action":{"type":"string","enum":["ADDITION","UPDATE"]},"messageState":{"oneOf":[{"$ref":"#/components/schemas/MessageStateResponse"},{"$ref":"#/components/schemas/FileStateResponse"},{"$ref":"#/components/schemas/IconStateResponse"},{"$ref":"#/components/schemas/ImageStateResponse"},{"$ref":"#/components/schemas/PreferenceStateResponse"},{"$ref":"#/components/schemas/TextStateResponse"}]}}},"MessageStateResponse":{"type":"object","discriminator":{"propertyName":"type"},"properties":{"sequenceNumber":{"type":"integer","format":"int64"},"postingId":{"type":"string","format":"uuid"},"sender":{"$ref":"#/components/schemas/PartnerResponse"},"messageType":{"type":"string"},"chatId":{"type":"string"},"createdAt":{"type":"string","format":"date-time"},"updatedAt":{"type":"string","format":"date-time"},"seenBy":{"type":"array","items":{"$ref":"#/components/schemas/PartnerResponse"},"uniqueItems":true},"type":{"type":"string"}},"required":["type"]},"PartnerResponse":{"type":"object","properties":{"id":{"type":"string","format":"uuid"},"name":{"type":"string"},"avatar":{"type":"string"}}},"PreferenceStateResponse":{"allOf":[{"$ref":"#/components/schemas/MessageStateResponse"}]},"TextStateResponse":{"allOf":[{"$ref":"#/components/schemas/MessageStateResponse"},{"type":"object","properties":{"content":{"type":"string"}}}]},"UserInfo":{"type":"object","properties":{"username":{"type":"string"},"name":{"type":"string"},"avatar":{"type":"string"},"id":{"type":"string","format":"uuid"}}},"ThemeResponse":{"type":"object","properties":{"id":{"type":"integer","format":"int32"},"background":{"type":"string"},"themeName":{"type":"string"}}},"RoomPresenceResponse":{"type":"object","properties":{"chatId":{"type":"string"},"at":{"type":"string","format":"date-time"}}},"PeopleResponse":{"type":"object","properties":{"id":{"type":"string","format":"uuid"},"username":{"type":"string"},"name":{"type":"string"},"avatar":{"type":"string"}}},"ConversationResponse":{"type":"object","properties":{"identifier":{"type":"string"},"roomName":{"type":"string"},"roomAvatar":{"type":"string"},"revisionNumber":{"type":"integer","format":"int64"},"recents":{"type":"array","items":{"oneOf":[{"$ref":"#/components/schemas/MessageStateResponse"},{"$ref":"#/components/schemas/FileStateResponse"},{"$ref":"#/components/schemas/IconStateResponse"},{"$ref":"#/components/schemas/ImageStateResponse"},{"$ref":"#/components/schemas/PreferenceStateResponse"},{"$ref":"#/components/schemas/TextStateResponse"}]}},"modifiedAt":{"type":"string","format":"date-time"}}},"MessageResponse":{"type":"object","properties":{"chatId":{"type":"string"},"sequenceNumber":{"type":"integer","format":"int64"},"content":{"type":"string"},"createdAt":{"type":"string","format":"date-time"}}},"BuddyResponse":{"type":"object","properties":{"userId":{"type":"string","format":"uuid"},"name":{"type":"string"},"avatar":{"type":"string"},"at":{"type":"string","format":"date-time"}}}},"examples":{"Validation":{"summary":"Validation failure","value":{"status":400,"title":"Validation failure","detail":"Email is invalid","type":"https://example.com/errors/validation"}},"NotFound":{"summary":"Resource not found","value":{"status":404,"title":"The requested resource was not found"}}},"securitySchemes":{"bearerAuth":{"type":"http","scheme":"bearer","bearerFormat":"JWT"},"basicAuth":{"type":"http","scheme":"basic"}}}}